Feat: 서버 나가기 API 수정
서버를 나가는 회원이 서버장일 경우에 대한 예외처리 코드 추가

+++ core/src/main/java/org/bookwoori/core/domain/server/facade/ServerFacade.java
@@ -136,6 +136,11 @@ public class ServerFacade {
     public void leaveServer(Long serverId) {
         Member member = memberService.getCurrentMember();
         Server server = serverService.getServerById(serverId);
+
+        if (serverMemberService.isOwner(member, server)) {
+            throw new CustomException(ErrorCode.DELEGATION_REQUIRED);
+        }
+
         serverMemberService.deleteServerMember(server, member);
     }
